requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.3
langchain-chroma==0.1.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
transformers==4.35.2
chromadb==0.4.15
numpy==1.24.3